feat: move around files more
