+++ lib/graph.js
@@ -175,7 +175,7 @@ const generateGraph = (result) => {
     svg.push(`  <text x="${x}" y="${y}" class="font s4" dominant-baseline="${baseline}" text-anchor="${anchor}">${benchmark.replace(/:/g, ' w/ ')}</text>`)
   })
 
-  svg.push(`  <text x="${graph.x + graph.w}" y="${vb.h - 2}" class="font s4 text" text-anchor="end">Tests run using node.js ${process.version}</text>`)
+  svg.push(`  <text x="${graph.x + graph.w}" y="${vb.h - 2}" class="font s4 text" text-anchor="end">Tests run with Node ${process.version}</text>`)
   svg.push('</svg>')
   return svg.join('\n')
 }
