+++ package.json
@@ -4,7 +4,6 @@
   "private": true,
   "dependencies": {
     "npm-package-arg": "^11.0.1",
-    "pacote": "^17.0.4",
     "semver": "^7.5.4",
     "yargs": "^17.7.2"
   },
@@ -45,7 +44,6 @@
   "devDependencies": {
     "@npmcli/eslint-config": "^4.0.2",
     "@npmcli/template-oss": "4.18.0",
-    "spawk": "^1.8.1",
     "tap": "^16.3.8",
     "which": "^4.0.0"
   }
